♻️ 32ZB9B runtime: decompose SGR contracts
