+++ .agentplane/tasks/202605290146-32ZB9B/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290146-32ZB9B`
+Title: SGR contracts decomposition
+Canonical task record: `.agentplane/tasks/202605290146-32ZB9B/README.md`
+
+## Summary
+
+SGR contracts decomposition
+
+Decompose packages/agentplane/src/runtime/sgr/contracts.ts by extracting contract types and shared validation primitives while preserving public SGR validation exports and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/runtime/sgr/contracts.ts by extracting contract types and shared validation primitives while preserving public SGR validation exports and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "SGR contracts decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T01:46:41.538Z
+- Branch: task/202605290146-32ZB9B/sgr-contracts-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
