+++ .agentplane/tasks/202605290146-32ZB9B/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 32ZB9B task: SGR contracts decomposition [202605290146-32ZB9B]
