+++ .agentplane/tasks/202605290146-32ZB9B/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T01:46:41.538Z
+
+## Task
+
+- Task: `202605290146-32ZB9B`
+- Title: SGR contracts decomposition
+- Status: DOING
+- Branch: `task/202605290146-32ZB9B/sgr-contracts-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290146-32ZB9B/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T01:46:41.538Z
+- Branch: task/202605290146-32ZB9B/sgr-contracts-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
